+++ interview/api/users-api/src/main/java/com/jonioliveira/users/domain/repository/UserTypeRepository.java
@@ -6,5 +6,5 @@ import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 import javax.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
-public class UserTypeRepository implements PanacheRepositoryBase<UserType, Long> {
+public class UserTypeRepository implements PanacheRepositoryBase<UserType, Integer> {
 }
